+++ HISTORY.md
@@ -1,6 +1,13 @@
 Release History
 ===============
 
+2.11.2 (26.02.2024)
+-------------------
+- Add Dictionaries to ResourceNames
+- Add DictionaryReader and DictionaryRegistrant roles.
+- Remove privileges Feeds, FeedsData and SearchFilters from roles FeedAdministrator и FeedDataReader
+- Remove privilege SearchFilters from role Searcher
+
 2.11.1 (2024-02-02)
 -------------------
 - Export PotentialDamage and RIR attributes